Refactor reverse_map into coarse_graining.py
