+++ mbuild/__init__.py
@@ -1,6 +1,5 @@
 from mbuild.box import Box
-from mbuild.coarse_graining import coarse_grain
-from mbuild.reverse_mapping import reverse_map
+from mbuild.coarse_graining import coarse_grain, reverse_map
 from mbuild.coordinate_transform import *
 from mbuild.compound import *
 from mbuild.pattern import *
